wip fix: avoid timeout when finishing the operator
In the past I observed that some development operation when shutting
down the operator on development environment, was taking always the
graceful time for the pod; looking at on detail I observed the pod was
not running the final process as PID 1 (it was the shell interpreter),
so the SIGTERM signal was not propagated.

This make the last command to run with 'exec' so the last command
replace the shell script, making it to receive the SIGTERM and avoiding
the graceful time.

TODO:

- Prepare a local environment.
- Check the operation in the local environment.
- Update this commit with the specific operation and how to replay it
  to properly check the change.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>